Decoupling of scanner and image facade
The image facade returns back in the body of the POST to the /checkimage
endpoint a URL encoded string containing the actual location of the
retrieved image. In most cases this will be a "file://" endpoint with
the actual location/file name used (as this *can* differ from the
PullSpec. The scanner uses that when calling the scan client.